feat: グローバルナビゲーション実装 - 全ページで統一ボトムナビ
## 主な変更
- 全ページで統一されたボトムナビゲーションを実装
- ダブルナビゲーション問題を解決
- MobileNavigationコンポーネントを新規作成

## UI改善
- ナビゲーションのアクセシビリティ向上（aria-label追加）
- アクティブ状態の視覚的強化（背景色、アニメーション）
- ホーム、カレンダー、タスク、OCR、設定の5タブ構成

## レイアウト統一
- app/layout.tsxでグローバルナビゲーション配置
- MobileLayoutコンポーネントからナビゲーション重複削除
- 各ページのヘッダーをMobileLayoutに統合
