+++ app/globals.css
@@ -46,15 +46,19 @@
 
 @layer components {
   .bottom-nav-item {
-    @apply flex flex-col items-center justify-center py-2 px-3 text-gray-600 transition-colors;
+    @apply flex flex-col items-center justify-center py-2 px-3 text-gray-500 transition-all duration-200 rounded-lg hover:bg-gray-50;
   }
   
   .bottom-nav-item.active {
-    @apply text-primary-600;
+    @apply text-primary-600 bg-primary-50;
   }
   
-  .safe-area-inset-bottom {
-    padding-bottom: env(safe-area-inset-bottom);
+  .bottom-nav-item:active {
+    @apply scale-95;
+  }
+  
+  .pb-safe {
+    padding-bottom: env(safe-area-inset-bottom, 0);
   }
   
   .card {
