+++ app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Noto_Sans_JP } from "next/font/google";
 import "./globals.css";
 import { DIProvider } from "@/components/providers/DIProvider";
+import MobileNavigation from "@/components/layout/MobileNavigation";
 
 const notoSansJP = Noto_Sans_JP({ 
   subsets: ["latin"],
@@ -32,7 +33,14 @@ export default function RootLayout({
     <html lang="ja" className="h-full">
       <body className={`${notoSansJP.variable} font-sans h-full`}>
         <DIProvider>
-          {children}
+          <div className="h-full flex flex-col">
+            <main className="flex-1 overflow-y-auto pb-16">
+              {children}
+            </main>
+            <nav className="fixed bottom-0 left-0 right-0 bg-white border-t border-gray-200 px-2 pb-safe">
+              <MobileNavigation />
+            </nav>
+          </div>
         </DIProvider>
       </body>
     </html>
