+++ app/page.tsx
@@ -59,7 +59,7 @@ export default function HomePage() {
   return (
     <div className="min-h-screen bg-gray-50">
       {/* ヘッダー */}
-      <header className="bg-white border-b border-gray-200">
+      <div className="bg-white border-b border-gray-200">
         <div className="max-w-6xl mx-auto px-4 py-6">
           <div className="text-center">
             <h1 className="text-2xl font-bold text-gray-900 mb-2">Sharendar</h1>
@@ -68,10 +68,10 @@ export default function HomePage() {
             </p>
           </div>
         </div>
-      </header>
+      </div>
 
       {/* メインコンテンツ */}
-      <main className="max-w-6xl mx-auto px-4 py-6 space-y-6">
+      <div className="max-w-6xl mx-auto px-4 py-6 space-y-6">
         {/* アラート */}
         {overdueTasks.length > 0 && (
           <div className="bg-red-50 border border-red-200 rounded-lg p-4">
@@ -355,7 +355,7 @@ export default function HomePage() {
             </Link>
           </div>
         </div>
-      </main>
+      </div>
     </div>
   );
 }
